+++ v6/js/modal.js
@@ -87,7 +87,7 @@
             this.container.classList.add("open");
             this.container.classList.add("ls-top-modal");
 
-            LS.Context.setTimeout(() => {
+            this.setTimeout(() => {
                 if(!this.isOpen || this.destroyed) return;
 
                 const focusable = this.container.querySelector("input, button, select, textarea, [tabindex]:not([tabindex='-1'])");
@@ -116,7 +116,7 @@
             this.container.classList.remove("ls-top-modal");
             LS.Stack.remove(this);
 
-            this.ctx.setTimeout(() => {
+            this.setTimeout(() => {
                 if(this.isOpen || this.destroyed) return;
 
                 if (LS.Stack.length === 0) {